+++ CHANGELOG.md
@@ -14,6 +14,10 @@ That's all, matcher will stop when detects next line started with `###` match
 * Added the option to join when a game is full (you will kick a player without perm)
 * Fixed issue that you could join full games before they started now proper full game
 permission check will occur and do the thing to allow or disallow you to join
+* Fixed the edit menu, now it will open
+* Fixed the arena signs
+* Fixed file creating for bungee.yml
+* Fixed VoteMenu not working
 * /bb randomjoin will now really join random arena not first one it find good to join
 
 ### 4.1.0-pre-1 Beta pre 1 (13/16.07.2019)